Refactored code to allow fetching of each rovers data.
Code has been updated to allow fetching of each rovers data.
If the rover has no data for the date, an exception is thrown and printed to let the user know.

The datasets are stored in raw JSON at their resepctive locations in S3.

I've also flattened the first level of array at this stage, compared to leaving it as extracted before.  This change will need to be propagated to the silver stage.